requirements.txt
streamlit
tensorflow-macos
Pillow
numpy